Rename app file, separate from init
